apiextensions: validate status.observedGeneration for custom resources
`status.observedGeneration` is useful to identify which `generation`
the status is valid for. This is mostly applicable when multiple
controllers are acting on the same object.

Right now, we already increment `metadata.Generation` on spec updates
and users can have  a `status.ObservedGeneration`.
However, we do not validate it. This commit adds validation for it.